Check for Microsoft packages

+++ android/setup.sh
@@ -3,9 +3,12 @@
 set -euox
 
 . /etc/os-release
-curl -sSk https://packages.microsoft.com/config/debian/$VERSION_ID/packages-microsoft-prod.deb -o m.deb
-sudo dpkg -i m.deb
-rm m.deb
+
+if [ ! "$(sudo dpkg --get-selections packages-microsoft-prod)" ]; then
+    curl -sSk https://packages.microsoft.com/config/debian/$VERSION_ID/packages-microsoft-prod.deb -o m.deb
+    sudo dpkg -i m.deb
+    rm m.deb
+fi
 
 sudo apt update && \
   sudo apt install -y \
